requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pytest
httpx
python-dotenv
pydantic[email]
requests